pyproject.toml
[project]
name = "clipboard-server"
version = "0.1.0"
description = "Clipboard sync API server"
requires-python = ">=3.11"
dependencies = [
    "requests>=2.32.4", # client
    "fastapi>=0.116.1",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.10.1",
    "pydantic[email]>=2.11.7",
    "python-jose>=3.5.0",
    "sqlalchemy>=2.0.41",
    "uvicorn[standard]>=0.35.0",
    "websockets>=15.0.1",
    "python-multipart>=0.0.20",
]
